requirements.txt
absl-py==2.1.0
alphalens-reloaded==0.4.5
appnope==0.1.4
asttokens==3.0.0
autopep8==2.3.1
backtrader==1.9.78.123
beautifulsoup4==4.12.3
black==24.10.0
Bottleneck==1.4.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.11
decorator==5.1.1
empyrical==0.5.5
empyrical-reloaded==0.5.11
exceptiongroup==1.2.2
executing==2.1.0
Farama-Notifications==0.0.4
filelock==3.4.1
finlab==1.2.20
flake8==7.1.1
fonttools==4.55.3
frozendict==2.4.6
fsspec==2024.12.0
grpcio==1.68.1
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lxml==5.3.0
lz4==4.3.3
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mpmath==1.3.0
multitasking==0.0.11
mypy-extensions==1.0.0
nbqa==1.9.1
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pandas-datareader==0.10.0
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
peewee==3.17.3
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pycodestyle==2.12.1
pyflakes==3.2.0
pyfolio==0.9.2
pyfolio-reloaded==0.9.8
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
Shimmy==2.0.0
six==1.17.0
soupsieve==2.6
stable_baselines3==2.4.0
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
ta-lib==0.5.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tokenize_rt==6.1.0
tomli==2.2.1
torch
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
xgboost==2.1.3
yfinance==0.2.51
zipp==3.21.0
